requirements.txt
# HealthMitra Scan – Python Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pydantic==2.9.0
ollama==0.4.0
aiofiles==24.1.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0

# Real AI Model Dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
ultralytics>=8.0.0
openai-whisper>=20231117
numpy>=1.24.0
